feat: add stardew valley mod
## Context

As a user, waiting for gold to smelt and trash to burn is time consuming.

## Decision

Modify machine game assets `(BC)13` furnace and `(BC)20` recyling machine to complete instantly upon adding an item.

## Consequences

The following items can be smelted instantly in the furnace:

- bouqet
- copper ore
- iron ore
- gold ore
- iridium ore
- radioactive ore
- quartz
- fire quartz

The following items can be recycled instantly in the recyling machine:

- trash
- driftwood
- broken glasses
- broken cd
- soggy newspaper